rename folderPullRequestManager file to folderRepositoryManager
