+++ src/github/pullRequestOverview.ts
@@ -14,7 +14,7 @@ import { writeFile, unlink } from 'fs';
 import Logger from '../common/logger';
 import { DescriptionNode } from '../view/treeNodes/descriptionNode';
 import { TreeNode, Revealable } from '../view/treeNodes/treeNode';
-import { FolderRepositoryManager } from './folderPullRequestManager';
+import { FolderRepositoryManager } from './folderRepositoryManager';
 import { PullRequestModel } from './pullRequestModel';
 import { TimelineEvent, ReviewEvent as CommonReviewEvent, isReviewEvent } from '../common/timelineEvent';
 import { IssueOverviewPanel, IRequestMessage } from './issueOverview';
